fixsed validete

+++ controllers/notices/removeNotice.js
@@ -3,14 +3,17 @@ const Notice = require("../../models/notice");
 
 const removeNotice = catchAsyncWrapper(async (req, res) => {
   const { noticeId } = req.params;
+  const notice = await Notice.findById(noticeId);
+  
 
-  const result = await Notice.findByIdAndRemove(noticeId);
-
-  if (!result) {
+  if (!notice) {
     throw httpError(404, "Not found");
   }
-
-  res.json({ message: "Announcement deleted" });
+  if (notice.owner.toString() !== req.user._id.toString()) {
+    throw httpError(400, "Bad request");
+  }
+  await Notice.findByIdAndRemove(noticeId);
+  res.json({ message: "Notice delete" });
 });
 
 module.exports = removeNotice;
