+++ controllers/pets/removePet.js
@@ -3,12 +3,17 @@ const Pet = require("../../models/pet");
 
 const removePet = catchAsyncWrapper(async (req, res) => {
   const { petId } = req.params;
-
-  const result = await Pet.findByIdAndRemove(petId);
-
-  if (!result) {
-    throw httpError(404, "Not found");
+  
+  const pet = await Pet.findById(petId);
+  
+  if (!pet) {
+    throw httpError(400, "Bad request");
   }
+
+  if (pet.owner.toString() !== req.user._id.toString()) {
+    throw httpError(400, "Bad request");
+  }
+  await Pet.findByIdAndRemove(petId);
 
   res.json({ message: "Pet deleted" });
 });
